Refactor code to fix issue with Interest Group renaming.
If an Interest Group gets renamed in Mailchimp, the sync scheduled job for
any CiviCRM groups set to sync with that Interest Group will fail. This
change allows an admin to resave the Group sync settings to get everything
working again. Prior to this change, you would have had to manually edit a
bunch of database records in the queue, which is no good.