More test cleanup work.
I added a phpunit.xml file to eventually allow us to easily run multiple
tests in a single civix test command. Once
https://github.com/totten/civix/pull/48 is accepted, this will be possible.

To prepare for that, I had to rename the Mailchimp API test class as the
existing naming scheme would have made PHPUnit think it needed to be tested,
which it shouldn't be as it's a mock class.