Add new sync log and admin error notification popups.
One of the things that makes syncing between two data systems problematic is
that when syncing fails, it can be difficult to determine when things stopped
working. I've added an admin accessible error log and notification popups
that are persistent and will warn an admin when syncing fails. The messages
will not go away until the admin selects to not display a message again.

This seems like the best way to avoid too much time passing after syncing
fails and should hopefully make fixing the issue easier.