Show an error message if there are no lists defined for a connected Mailchimp account.
If there was no error connecting to Mailchimp to retrieve Mailchimp Lists,
it's difficult to see why the Mailchimp Sync Settings when editing a Group
are not showing Mailchimp Lists. The reason this happens is because there are
no lists defined in the Mailchimp account, so I've added a messsage to that
effect when there are no lists defined.