Do not die when syncing group with Contacts with no email.
We need to allow CiviCRM groups to sync to Mailchimp and not throw a fatal
error if there is a Contact in that group without a valid email to sync to
Mailchimp. This update takes care of that and also adds a log message when a
Force Sync is performed which informs an admin of how many contacts were
skipped.

I also added a new setting to allow for marking a Contact as No Bulk Emails
if they are unsubscribed through Mailchimp. Enabling this setting allows us
to deal with Smart Groups using a workaround where the admin can create a
one-off group based on the Smart Group for sending things like monthly emails
to current members. In that use-case, if a current member unsubscribes
through Mailchimp, they will no longer recieve an email even though they will
be in the next one-off group, since they will have been marked as No Bulk
Emails.