Allow admins to clear errored items from the sync queue.
Most of the time, when there is an error in the sync queue, it is fixable by
admins by manually editing a record in either CiviCRM or Mailchimp. However,
we've come across a specific situation where this can't be done.

Mailchimp tracks an "invalid email" list which includes email addresses that
they refuse to add to a list. If a Contact in CiviCRM has one of those emails
and a queue item is created to subscribe them to a Mailchimp list, the
subscribe API request gets an error response from Mailchimp. At this point,
if we wanted, we could just delete the queue item if we get this specific
kind of error back from Mailchimp. However, this does not solve another
problem.

If that invalid email is in a CiviCRM group and then the group is configured
to sync to Mailchimp, if the user is then removed from the CiviCRM group, a
queue item is created to unsubscribe them from the connected Mailchimp list.
However, since Mailchimp does not allow subscribing the invalid email address
in the first place, we can't go back to Mailchimp and temporarily subscribe
the email addresss, which is how we would normally un-stick the queue.

So, we need to provide a mechanism for allowing admins to determine that they
would no longer like to try syncing an item in the queue.